fpt/requirements.txt
# requirements.txt
tensorflow>=2.12.0
requests>=2.28.1
grpcio>=1.48.1
grpcio-tools>=1.48.1
opencv-python>=4.6.0.66
numpy>=1.23.2
ultralytics>=8.0.39
